requirements.txt
google-generativeai>=0.2.2
pillow>=10.2.0
python-docx>=1.0.0
PyPDF2>=3.0.1
tqdm>=4.66.2
watchdog>=3.0.0
exifread>=3.0.0 